Moved calculations from StatsService to StatsHelper
By adding a StatsHelper and moving all the calculations there (leaving only the business logic in StatsService), we keep StatsService is focused.
It improves reusability, and makes testing easier.
