+++ src/main/java/com/selekode/topaz/repository/JournalRepository.java
@@ -1,4 +1,4 @@
-package com.selekode.repository;
+package com.selekode.topaz.repository;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
